requirements.txt
packaging==24.2
pillow==11.0.0
pyinstaller-hooks-contrib==2024.10
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
setuptools==75.6.0
zipp==3.21.0
ping3==4.0.4  # Add this line